+++ core/ui/src/commonMain/composeResources/drawable-hdpi/placeholder_speaker.png
