+++ core/ui/src/commonMain/composeResources/drawable-xhdpi/placeholder_speaker.png
