+++ core/ui/src/commonMain/composeResources/drawable-xxhdpi/placeholder_speaker.png
