+++ core/ui/src/commonMain/composeResources/drawable-xxxhdpi/img_cat_archer.png
