+++ core/ui/src/commonMain/composeResources/drawable-xxxhdpi/img_cat_wizard.png
