+++ core/ui/src/commonMain/composeResources/drawable-xxxhdpi/placeholder_speaker.png
