+++ gradle/libs.versions.toml
@@ -154,6 +154,7 @@ compose-navigation = { group = "org.jetbrains.androidx.navigation", name = "navi
 androidx-runner = { group = "androidx.test", name = "runner", version.ref = "androidxTestRunner" }
 compose-compiler-gradle-plugin = { module = "org.jetbrains.kotlin:compose-compiler-gradle-plugin", version.ref = "kotlin" }
 compose-gradle-plugin = { module = "org.jetbrains.compose:compose-gradle-plugin", version.ref = "composePlugin" }
+compose-component-resource = { group = "org.jetbrains.compose.components", name = "components-resources", version.ref = "composePlugin" }
 ksp-gradle-plugin = { module = "com.google.devtools.ksp:com.google.devtools.ksp.gradle.plugin", version.ref = "ksp" }
 
 hilt-core = { group = "com.google.dagger", name = "hilt-core", version.ref = "hilt" }
